py/requirements.txt
protobuf==3.13.0
psycopg2-binary==2.8.6
six==1.15.0
